+++ demisto_sdk/commands/common/constants.py
@@ -197,6 +197,7 @@ RN_HEADER_BY_FILE_TYPE = {
     FileType.XSIAM_REPORT: 'XSIAM Reports',
     FileType.TRIGGER: 'Triggers Recommendations',  # https://github.com/demisto/etc/issues/48153#issuecomment-1111988526
     FileType.WIZARD: 'Wizards',
+    FileType.XDRC_TEMPLATE: 'XDRC Templates',
 }
 
 ENTITY_TYPE_TO_DIR = {